spotify-extract/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

logging
requests
azure-functions
datetime
azure-storage-blob